v2/chapter4/requirements.txt
mecab-python3 == 1.0.1
unidic-lite == 1.0.7
matplotlib
ipython
ipykernel